% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_PMD.R
\name{characterize_invariance_property}
\alias{characterize_invariance_property}
\title{characterize_invariance_property}
\usage{
characterize_invariance_property(
  directory = "",
  num_clust_per_batch_range = seq(4, 17, by = 4),
  b1_size_vect = c(10000, 10000, 1000, 250, 250, 100),
  b2_size_vect = c(10000, 1000, 1000, 2000, 250, 100),
  percent_difference = c(0, 0.25, 0.5, 0.75, 1),
  iterations = 10,
  expand_b2_only = FALSE
)
}
\arguments{
\item{directory}{the directory for putting the output image. This isn't needed though;
we'll skip the plotting in that case.}

\item{num_clust_per_batch_range}{a vector of number of clusters (as integers)}

\item{b1_size_vect}{a vector denoting the size for the first batch for the given iteration}

\item{b2_size_vect}{a vector denoting the size for the second batch for the given iteration}

\item{percent_difference}{a floating point betwee 0. and 1. that corresponds to the 
percentage difference between the two batches}

\item{iterations}{an integer that}

\item{expand_b2_only}{a TRUE/FALSE boolean operator that dictates the mechanism of simulation in terms of what the 
\code{num_clust_per_batch_range} argument operates on. When FALSE, both batches
are expanded with increasing numbers of groups & what determins \code{percent_difference} is the number
of shared clusters. When TRUE, batch1 has all its cells allocated to a single cluster & batch 2
has the \code{percent_difference} percent of cells allocated to that same cluster as batch 1, and the rest
of its cells are partitioned equally across the remaining clusters.}
}
\value{
Null
}
\description{
\code{characterize_invariance_property} Run the PMD characterization over a range of power conditions.
}
\examples{
   pmd_bench_table_combined <- characterize_invariance_property(b1_size_vect = c(250, 250),
                                                               b2_size_vect = c(500, 1000),
                                                               iterations = 3)
}
