% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_PMD.R
\name{do_full_pmd_characterization}
\alias{do_full_pmd_characterization}
\title{do_full_pmd_characterization}
\usage{
do_full_pmd_characterization(directory = "")
}
\arguments{
\item{directory}{the directory for putting the output image}
}
\value{
Null
}
\description{
\code{do_full_pmd_characterization} Run the PMD characterization over a range of power conditions.
}
\examples{
   do_full_pmd_characterization()
}
