% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_cont_table}
\alias{get_cont_table}
\title{get_cont_table}
\usage{
get_cont_table(x, y)
}
\arguments{
\item{x}{a vector that can be interpreted as a factor. Should be clusters}

\item{y}{a vector that can be interpreted as a factor. Should be batches}
}
\value{
a contingency matrix with x quantified in rows, y quantified in columns
}
\description{
\code{get_cont_table} Gets a congingency table that quantifies two factors against each other, assuming that they are in the same order.
}
\examples{
   x <- rep( c( rep(1, 50), rep(2, 50) ), 2)
   y <- c(rep(1, 100), rep(2, 100))
   cont_table <- get_cont_table(x, y)
}
