% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_expected}
\alias{get_expected}
\title{get_expected}
\usage{
get_expected(cont_table)
}
\arguments{
\item{cont_table}{The observed contingency table}
}
\value{
a matrix representing the 'expected' values if there were no pattern
}
\description{
\code{get_expected} Gets the expected matrix, as with a Chi Squared test.
}
\examples{
   cont_table <- matrix(c(50,100,30,80,90,50),nrow = 3, ncol = 2)
   get_expected(cont_table)
}
