% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_percent_max_diff}
\alias{get_percent_max_diff}
\title{get_percent_max_diff}
\usage{
get_percent_max_diff(group1_labs, group2_labs)
}
\arguments{
\item{group1_labs}{a vector of labels for the batch that each sample (or cell for scRNAseq) came from. Can be a string or factor.}

\item{group2_labs}{a vector of labels for each cluster that each sample belongs to. Can be a string or factor.}
}
\value{
a floating point number of the percent maximum difference
}
\description{
\code{get_percent_max_diff} Gets the Percent Maximum Difference (PMD) for clustering results relative to batch.
}
\examples{
   batch <- rep(seq_len(2),each=100)
   clusters <- rep(rep(seq_len(2),each=50),2)
   pmd_num <- get_percent_max_diff(batch, clusters)
}
