% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_percent_max_diff_from_tables}
\alias{get_percent_max_diff_from_tables}
\title{get_percent_max_diff_from_tables}
\usage{
get_percent_max_diff_from_tables(groups1_table, groups2_table)
}
\arguments{
\item{groups1_table}{a data frame in which the first column is the cell IDs, and second column is the cluster that cell belongs to.}

\item{groups2_table}{a data frame of the same format as above, but for the second dataset.}
}
\value{
pmd
}
\description{
\code{get_percent_max_diff_from_tables} calculates the percent max difference for two batches 
                that have the same cells. This can be useful if you want to do something to the same dataset &
                look at how it changes the results by PMD.
}
