% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_pmd_null_vect}
\alias{get_pmd_null_vect}
\title{get_pmd_null_vect}
\usage{
get_pmd_null_vect(expected_mat, num_sim = 10000)
}
\arguments{
\item{expected_mat}{A matrix that represents a contingency matrix of clusters (rows) and batch (cols).}

\item{num_sim}{for generating a null distribution in calculating a p-value, the number of simulations to run.}
}
\value{
a vector of floating points that represent a null distribution that matches your input for percent maximum difference
}
\description{
\code{get_pmd_null_vect}
}
\examples{
   dummy_mat <- matrix(c(500, 250, 100, 1000, 500, 200),nrow = 3, ncol=2)
   pmd_null <- get_pmd_null_vect(dummy_mat)
}
