% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_random_sample_cluster}
\alias{get_random_sample_cluster}
\title{get_random_sample_cluster}
\usage{
get_random_sample_cluster(prob_vect, num_samples)
}
\arguments{
\item{prob_vect}{a vector of probabilities that sum up to 1, that represent the relative abundance of each clusters that we'll be sampling from.}

\item{num_samples}{the number of samples to simulate.}
}
\value{
a vector of simulated cluster labels of length \code{num_samples} based on the probility vector that codes the relative abundance of each cluster.
}
\description{
\code{get_random_sample_cluster}
}
\examples{
   prob_vect <- c(.25, .25, .5)
   sim_clusters <- get_random_sample_cluster(prob_vect, 1000)
}
