% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{pmd}
\alias{pmd}
\title{Percent Maximum Difference}
\usage{
pmd(batch_labels, cluster_labels, num_sim = 10000)
}
\arguments{
\item{batch_labels}{a vector of labels for the batch that each sample (or cell for scRNAseq) came from. Can be a string or factor.}

\item{cluster_labels}{a vector of labels for each cluster that each sample belongs to. Can be a string or factor.}

\item{num_sim}{for generating a null distribution in calculating a p-value, the number of simulations to run.}
}
\value{
list object
   \enumerate{
   \item \code{cont_table} - blah blah
   \item \code{chi} A list containing all of the results from a traditional Chi Sqr
          as with the chisq.test function.
   \item \code{pmd} The percent maximum difference (pmd) of the input dataset.
   \item \code{pmd_null} Simulations of the null distribution of PMDs using the 
          observed global percentage of cluster abundances across all batches 
          with the observed batch sizes to match the input.
          Note that this will approach zero, but due to random sampling, typically will
          never actually get there. That's what makes having this null background useful.
   \item \code{pmd_z} The Z-statistic for how different the observed PMD is from the 
          distribution of null PMDs. However, in many cases, the
          null background distribution of PMDs is a skewed Poisson or beta-like distribution
          so a Z-statistic is less meaningful in that case. The returned \code{p.value} 
          however is generated as an empirical p-value as measured against the simulations.
    \item \code{p.value} - Significance for whether or not batches are different in their cellular composition.
          This is determined through the generating \code{num_sim} null distributions, 
          and emperically measuring the number of times the observed PMD was greater 
          than the PMDs generated from the null distribution.
          Low p-values indicate that the batches are indeed different from each other.
   }
}
\description{
\code{pmd} Percent Maximum Difference provides a metric that enables the measurement
                        of how similar (PMD near 0) or different (PMD near 1) two batches are
                        based just on clustering results. This can be viewed as an alternative
                        to a Chi-squared type analysis based on a contingency table in which 
                        the batches are noted in columns, and the abundance of cells in each cluster
                        is noted in rows. PMD advances on Chi-squared and Fishers exact tests
                        because its results are actually not dependent on the number of clusters found.
                        It does this by first calculating the hypothetical maximum possible asymmetry 
                        between your batches. (i.e.: if each batch was only comprised of clusters 
                        that only ever appeared in that batch and there were no clusters that appeared 
                        in more than one batch). Then it calculates the observed asymmetry in your 
                        real results, then returning the percentage of the observed asymmetry
                        relative to the maximum possible asymmetry. This gives a lower bound of zero,
                        where the relative abundance of each cluster was spot-on the same between
                        all of the batches. This also gives an upper bound of one, in which 
                        the observed clustering results were fully asymmetric across batch; 
                        or in other words, each cluster was fully batch specific.
}
\examples{
   batch <- rep(seq_len(2),each=100)
   clusters <- rep(rep(seq_len(2),each=50),2)
   pmd_res <- pmd(batch, clusters)
}
