% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{pmd_from_cont_table}
\alias{pmd_from_cont_table}
\title{Percent Maximum Difference from a contingency table}
\usage{
pmd_from_cont_table(cont_table, num_sim = 1000)
}
\arguments{
\item{cont_table}{As an alternative to passing in batch and cluster labels, you can also just pass in the contingency table.
Note that the clusters MUST be in the rows, and the batches must be the columns of this matrix}

\item{num_sim}{for generating a null distribution in calculating a p-value, the number of simulations to run.}
}
\value{
list object
   \enumerate{
   \item \code{cont_table} The contingency table of clusters (rows) and batches (columns)
   \item \code{expected} The expected matrix, as with a Chi-square test.
   \item \code{pmd_null} Simulations of the null distribution of PMDs using the 
          observed global percentage of cluster abundances across all batches 
          with the observed batch sizes to match the input.
          Note that this will approach zero, but due to random sampling, typically will
          never actually get there. That's what makes having this null background useful.
          Note that these are Raw PMD, not adjusted.
    \item \code{pmd_null_lambda} The lambda value of the null distributions Poisson fit. 
          This is used in calculating the final PMD value compared to the raw PMD value.
    \item \code{pmd_raw} The original, Raw PMD, before power correction
    \item \code{p.value} - Significance for whether or not batches are different in their cellular composition.
          This is determined through the generating \code{num_sim} null distributions, 
          and emperically measuring the number of times the observed PMD was greater 
          than the PMDs generated from the null distribution.
          Low p-values indicate that the batches are indeed different from each other.
    \item \code{pmd} The percent maximum difference (pmd) of the input dataset.
   }
}
\description{
\code{pmd_from_cont_table} Percent Maximum Difference provides a metric that enables the measurement
                        of how similar (PMD near 0) or different (PMD near 1) two batches are
                        based just on clustering results. This can be viewed as an alternative
                        to a Chi-squared type analysis based on a contingency table in which 
                        the batches are noted in columns, and the abundance of cells in each cluster
                        is noted in rows. PMD advances on Chi-squared and Fishers exact tests
                        because its results are actually not dependent on the number of clusters found.
                        It does this by first calculating the hypothetical maximum possible asymmetry 
                        between your batches. (i.e.: if each batch was only comprised of clusters 
                        that only ever appeared in that batch and there were no clusters that appeared 
                        in more than one batch). Then it calculates the observed asymmetry in your 
                        real results, then returning the percentage of the observed asymmetry
                        relative to the maximum possible asymmetry. This gives a lower bound of zero,
                        where the relative abundance of each cluster was spot-on the same between
                        all of the batches. This also gives an upper bound of one, in which 
                        the observed clustering results were fully asymmetric across batch; 
                        or in other words, each cluster was fully batch specific.
}
\examples{
   ## generate the contingency table
   cont_table <- matrix(c(50,100,30,80,90,50),nrow = 3, ncol = 2)
   ## Note that the clusters are in rows, and the batches are in columns 
   ## of this matrix. This is important!
   pmd_res <- pmd_from_cont_table(cont_table)
}
