% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{pmd_posthoc}
\alias{pmd_posthoc}
\title{Percent Maximum Difference}
\usage{
pmd_posthoc(pmd_result, num_sim = NULL, adjust_method = "BH")
}
\arguments{
\item{pmd_result}{The result of the \code{pmd} function.}

\item{num_sim}{the number of simulations to perform for both the lambda adjustment and p-value calculations}

\item{adjust_method}{the method of p-value adjustment}
}
\value{
list object
   \enumerate{
    \item \code{pmd_raw_table} 
    \item \code{pmd_table} 
    \item \code{p.value_table} 
    \item \code{p.value_table_adjusted} 
   }
}
\description{
\code{pmd_posthoc} The PMD posthoc does a pairwise comparison of all the columns to each other
                                This allows you to figure out which of the columns were actually different 
                                from each other. This also gives a quantitative metric for exactly how similar
                                or different the columns are from each other based on their row-wise composition.
}
\examples{
   ## first using batch and cluster labels
   batch_labels <- paste("batch",rep(seq_len(3),each=100))
   cluster_labels <- paste("cluster",c(rep(rep(seq_len(2),each=50),2),rep(seq_len(2),each=50)+1))
   pmd_res <- pmd(batch_labels, cluster_labels)
   ## This is what the cluster/batch contingency table looks like:
   print(pmd_res$cont_table)
   ## Which makes the percent maximum difference ~ 1/3rd
   print(pmd_res$pmd_raw)
   ## in a real-world sceanrio, the data wouldn't be this clean beacuse of noise via Poisson sampling
   ## so the corrected PMD is actually a bit less:
   print(pmd_res$pmd)
   ## Using the pmd_posthoc function, we'll be able to figure out exactly which batch(es) are causing the 1/3rd asymmetry
   pmd_pairwise <- pmd_posthoc(pmd_res)
}
