% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_PMD.R
\name{characterize_pmd}
\alias{characterize_pmd}
\title{characterize_pmd}
\usage{
characterize_pmd(
  num_cells_b1 = 1000,
  num_cells_b2 = 1000,
  num_clust_shared = seq(from = 0, to = 10),
  iters = 10,
  pmd_2 = FALSE
)
}
\arguments{
\item{num_cells_b1}{The number of cells in batch 1}

\item{num_cells_b2}{The number of cells in batch 2}

\item{num_clust_shared}{the number of clusters that are shared across the two batches}

\item{iters}{the number of iterations for each condition}

\item{pmd_2}{A boolean argument from when we had thought about a standardized version of PMD that used squares/sqrts instead of absolute values. It worked much worse, so don't bother using this arg.}
}
\value{
list object
   \enumerate{
   \item \code{out_df}  A dataframe that has the results of the characterization. Also returns a vector 
   \item \code{null_df} A dataframe with the vector of the PMDs generated from the null distribution of the input.
   }
}
\description{
\code{characterize_pmd} Run the PMD characterization over a range of power conditions.
}
\examples{
   characterize_pmd()
}
