% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_PMD.R
\name{characterize_pmd}
\alias{characterize_pmd}
\title{characterize_pmd}
\usage{
characterize_pmd(
  num_cells_b1 = 1000,
  num_cells_b2 = 1000,
  num_clust_shared = 0:10,
  iters = 10
)
}
\arguments{
\item{num_cells_b1}{The number of cells in batch 1}

\item{num_cells_b2}{The number of cells in batch 2}

\item{num_clust_shared}{the number of clusters that are shared across the two batches}

\item{iters}{the number of iterations for each condition}
}
\value{
Null
}
\description{
\code{characterize_pmd} Run the PMD characterization over a range of power conditions.
}
\examples{
   characterize_pmd()
}
