% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_PMD.R
\name{do_full_pmd_characterization}
\alias{do_full_pmd_characterization}
\title{do_full_pmd_characterization}
\usage{
do_full_pmd_characterization(
  directory = "",
  b1_size_vect = c(10000, 10000, 1000, 250, 250, 100),
  b2_size_vect = c(10000, 1000, 1000, 2000, 250, 100)
)
}
\arguments{
\item{directory}{the directory for putting the output image. This isn't needed though;
we'll skip the plotting in that case.}

\item{b1_size_vect}{a vector denoting the size for the first batch for the given iteration}

\item{b2_size_vect}{a vector denoting the size for the second batch for the given iteration}
}
\value{
Null
}
\description{
\code{do_full_pmd_characterization} Run the PMD characterization over a range of power conditions.
}
\examples{
   pmd_bench_table_combined <- do_full_pmd_characterization(b1_size_vect = c(250, 250),
                                                            b2_size_vect = c(500, 1000))
}
