% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMaximumDifference.R
\name{get_percent_max_diff_from_tables}
\alias{get_percent_max_diff_from_tables}
\title{get_percent_max_diff_from_tables}
\usage{
get_percent_max_diff_from_tables(groups1_table, groups2_table)
}
\arguments{
\item{groups1_table}{a data frame in which the first column is the cell IDs, and second column is the cluster that cell belongs to.}

\item{groups2_table}{a data frame of the same format as above, but for the second dataset.}
}
\value{
pmd
}
\description{
\code{get_percent_max_diff_from_tables} Calculates PMD from two input dataframes. The first data frame
                has cell IDs in first column, and batch in the second column. The second
                dataframe has the Cell IDs and then the clusters
}
\examples{
      groups1_table = data.frame(cell = paste("cell",1:10),
                                 batch = c(rep(1,5),rep(2,5)))
      groups2_table = data.frame(cell = paste("cell",1:10),
                                 clust = c(rep(1,5),rep(2,5)))
      get_percent_max_diff_from_tables(groups1_table, groups2_table)
}
